Restructure repo to add seeding folder
