+++ db/seeding/dataGeneration.js
@@ -1,5 +1,5 @@
-const db = require('./index.js');
-const S3 = require('./s3FileUpload.js');
+//const db = require('./index.js');
+const S3 = require('../s3FileUpload.js');
 const fs = require('fs');
 const csvWriter = require('csv-write-stream');
 const writer = csvWriter({ headers: ['item_id', 'image_url']});
@@ -10,7 +10,7 @@ const randomNumber = (min, max) => {
 }
 
 const writeCSV =  async (numberOfItemsToSeed) => {
-  writer.pipe(fs.createWriteStream('./db/seedData.csv'));
+  writer.pipe(fs.createWriteStream('./db/seeding/seedData.csv'));
   let urls = await S3.getUrlsS3();
 
   for (let i = 0; i < numberOfItemsToSeed; i++){
@@ -20,17 +20,13 @@ const writeCSV =  async (numberOfItemsToSeed) => {
       let randomIndex = randomNumber(0, urls.length);
       let currentValue = urls[randomIndex];
       let values = [itemId, currentValue];
-      //console.log(values);
       writer.write(values);
-      // db.query('INSERT INTO images (item_id, image_url) VALUES ($1, $2)', values, (err, data) => {
-        //   if (err) { console.error(err) }
-        // });
       }
     }
     writer.end();
 }
 
-//input desired number
+//input desired number as parameter
 writeCSV();
