+++ db/seeding/seedPSQL.js
@@ -1,4 +1,4 @@
-const db = require('./index.js');
+const db = require('../index.js');
 
 const insertData = async () => {
   let csv = __dirname + '/seedData.csv';
