+++ package.json
@@ -7,7 +7,8 @@
     "test": "jest",
     "start": "nodemon server",
     "react-dev": "webpack --watch",
-    "db:setup": "psql < db/schema.sql && node db/seed.js"
+    "db:setup": "psql < db/schema.sql && node db/seeding/dataGeneration.js",
+    "db:seed": "node db/seeding/seedData.js"
   },
   "repository": {
     "type": "git",
